chore: enable fetch to be overriden by optional params to buildFetch based on review feedback
