--- packages/network/__tests__/index-test.ts
@@ -1,5 +1,0 @@
-import { describe, it } from 'vitest';
-
-describe('@data-eden/network', function() {
-  it('can run tests', function() {});
-});
